requirements.txt
git-filter-repo
django
django-filter
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
pymysql
mysqlclient
mysql-connector-python
Faker
requests
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client


